# config/swaggerOptions.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "NodeJs Bootcamp Documentation API",
    "version": "1.0.0",
    "contact": {
      "name": "a bai",
      "email": "1121176166@qq.com",
      "url": "http://baboonbutts.nft"
    }
  },
  "servers": [{
    "description": "Local Server",
    "url": "http://localhost:3001"
  }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
  "security": [{
    "bearerAuth": []
  }],
  "paths": {
    "/user/register": {
      "post": {
        "tags": ["users"],
        "description": "Register User",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "default": "abai"
                  },
                  "lastName": {
                    "type": "string",
                    "default": "haha"
                  },
                  "address": {
                    "type": "string",
                    "default": "河南省洛阳市伊川县见新一届9号"
                  },
                  "city": {
                    "type": "string",
                    "default": "洛阳市"
                  },
                  "state": {
                    "type": "string",
                    "default": "中国"
                  },
                  "zipCode": {
                    "type": "string",
                    "default": "47130"
                  },
                  "email": {
                    "type": "string",
                    "default": "1121176166@qq.com"
                  },
                  "password": {
                    "type": "string",
                    "default": "a123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error"
          },
          "501": {
            "description": "Error"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["users"],
        "description": "Login User",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "default": "1121176166666@qq.com"
                  },
                  "password": {
                    "type": "string",
                    "default": "a123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error"
          },
          "501": {
            "description": "Error"
          }
        }
      }
    },
...
  }
}